Issue #3933: Split and Organize Checkstyle inputs by Test for NoLineWrap
